--- frontend/src/data/theads.ts
@@ -1,37 +1,0 @@
-import { IThead } from './interfaces';
-
-export const userTableHead: IThead[] = [
-  {
-    name: 'ID'
-  },
-  {
-    name: 'Name'
-  },
-  {
-    name: 'Role'
-  },
-  {
-    name: 'Email'
-  },
-  {
-    name: 'Date Created'
-  },
-]
-
-export const categoryTableHead: IThead[] = [
-  {
-    name: 'ID'
-  },
-  {
-    name: 'Title'
-  },
-  {
-    name: 'Description'
-  },
-  {
-    name: 'Date Created'
-  },
-  {
-    name: 'Actions'
-  }
-]
